Add CI build as pipeline resource in E2E-Checkin pipeline (#6635)
My recent update (a674bad) to unify the image and package builds introduced a regression. I removed the pipeline resources defined in e2e-checkin.yaml because I thought they were unused (my runs passed without them). But it turns out they _are_ needed in cases where either images or packages aren't built inline. This can happen when, for example, the PR that triggers the pipeline only made changes to one and not the other. The pipeline resources serve as the fallback source for either images or packages in this case.

## Azure IoT Edge PR checklist: