Use preloadable temporary module in ValidateMetrics (#6648)
The `ValidateMetrics` end-to-end test deploys a module to the test device then removes it in order to bump a particular metric. It doesn't matter which module is used; we just need to add something then remove it. Currently, we've hardcoded `mcr.microsoft.com/azureiotedge-simulated-temperature-sensor:1.0`. Because that isn't one of the images listed in context.json, our test pipeline doesn't preload it on slow test devices, which means it has to be downloaded during the test, which eats into the test's timeout duration.

The `ValidateMetrics` test is consistently timing out on Raspberry Pi. Changing the test to use the tempSensor image from context.json _should_ speed up the test and prevent the timeouts.

I ran the end-to-end pipeline against these changes and `ValidateMetrics` passed on Raspberry Pi.