Upgrade to Rust 1.63 (#6674)
This is our monthly toolchain upgrade to the latest stable version. Since the primary artifacts of this repository are binaries, we want to be tracking Rust compiler releases closely in the event a stdlib vulnerability is found. We do not pin to "stable", however, since that has caused pipeline breakage when some code patterns are made illegal (like in the upgrade from 1.47 to 1.48 [^0]) or, more often, clippy lints are added.

- `clippy::borrow_deref_ref`: remove unnecessary uses of `&*`
- `clippy::derive_partial_eq_without_eq`: derive `Eq` where possible

## Azure IoT Edge PR checklist: