Update Azure Functions sample's base image (#6698)
The azure-functions/dotnet base image we're using in our Azure Functions sample (edge-modules/functions/samples) is not maintained (it is about 9 months old). The internal images we're using in our tests have security vulnerabilities. This change updates to a base image that is regularly serviced by the Azure Functions team.

To test, I built a new azureiotedge-functions-filter image, ran the end-to-end tests against it, and confirmed they passed.

## Azure IoT Edge PR checklist: