Set connection and response timeouts to 5min (#6716)
Set Connection Timeout and Response Timeout to max values of 5 minutes in DirectMethodSender test module used by the ModuleToModuleDirectMethod method. This should reduce the failure rate of the test on ARM32v7 platforms where direct methods can take a long time to complete.

To test this change a CI Build pipeline was run from this branch and then an E2E pipeline was run using the resulting artifacts.