Move metrics controller into main branch (#6713)
The Metrics Collector module, which works with any version of IoT Edge, has been maintained and released out of the `release/1.1` branch, which is set to go out of support in two months. This change moves the module's source code and test/release pipelines to the `main` branch, where it can continue to be supported independent of any specific IoT Edge version.

Most of the _changes_ in this PR are in the test/release pipelines and scripts. Under the source code directory `edge-modules/metrics-collector`, everything is an exact copy from the `release/1.1` branch except:
- `edge-modules/metrics-collector/src/Microsoft.Azure.Devices.Edge.Azure.Monitor.csproj`
  * Upgrade from .NET Core 3.1 to .NET 6.0
  * Upgrade the Device SDK reference to match the version used throughout the `main` branch
- `edge-modules/metrics-collector/docker/linux/**/Dockerfile`
  * Base Docker images on Alpine
  * Remove Windows image
- `edge-modules/metrics-collector/docker/manifest.yaml.template`
  * Remove Windows manifest entry

To test, I verified that:
- the CI Build pipeline builds the module's source code and Docker images
- the end-to-end test pipeline successfully runs the MetricsCollector test
- the metrics collector release pipeline is functional (although we won't officially release it from this branch until after 1.1 ends)