Add Mariner 2.0-stable back to 1.4 Branch (#6727)
This PR will re-enable the release/1.4 Mariner package to be built using Mariner 2.0-stable. For This build to work the required changes to fix the iot-identity-service package for mariner must also be ingested, which is why the hash has been updated to reflect the most recent version of iot-identity-service release/1.4

The release/1.4 branch of iot-identity-service is being used for the mariner package's vendoring step. 

## Azure IoT Edge PR checklist: