Add E2E and LH Tests for Mariner in Main Branch (#6539)
This PR enables Running E2E and LH test to validate the packages created for mariner. By default, these pipelines only worked with .deb packages, so most of the worked for this PR involved removing hardcoded paths that relied on the .deb extension and adding additional logic for utilizing .rpm files. When possible, I tried to encapsulate the logic for interacting with specific package manger in their own classes.

## Azure IoT Edge PR checklist: