Add a pipeline that detects base image updates (#6785)
Cherry-pick 6e3ac3cf117e9f1b1de5b1e4e4014214923f189b.

I had to make slight tweaks to accommodate the Metrics Collector module, which the `release/1.4` branch doesn't have.

Most of the changes aren't strictly needed in `main` because the core modules release out of `release/1.4`, but some changes are needed for automated refresh of the Metrics Collector module (which will come in later PR).

To test, I exercised the pipelines that run in `main` today and require any of the templates/scripts that were changed in this PR:
- .NET Checkin (buildBranch.sh)
- E2E Checkin (build-images.yaml, build-rocksdb.yaml, image-linux.yaml, buildBranch.sh)
- CI Build (build-images.yaml, build-rocksdb.yaml, image-linux.yaml, buildBranch.sh)