Move remaining pipelines to 1ES agents (#6819)
We still have a few Azure DevOps pipelines in our codebase the use Microsoft-hosted agents, and most of them are using "ubuntu18.04" which will be deprecated soon. This change updates the remaining references to Ubuntu 20.04 1ES agents.

To test, I ran the affected pipelines against the PR (except the release pipeline, but it is very similar to the edgelet part of the CI build, which already uses Ubuntu 20.04 1ES agents) and confirmed that they succeed.

## Azure IoT Edge PR checklist: