Fix jq logic in image refresh pipeline (#6818)
I was expecting our 1.4 core Docker images to automatically refresh shortly after the December .NET 6.0 update, but it didn't happen. I investigated and discovered a bug in the jq logic that decides whether any of the detected updates apply to the 1.4 core images. The flaw went unnoticed in our first automatic refresh (1.4.4) because the diagnostics image happened to have the same version (1.4.3) as the other core images. But the diagnostics image is versioned differently (it always matches edgelet) so it will not always match the other core image versions. That's what happened this time--agent, hub, and temp sensor are at 1.4.4, while diagnostics is at 1.4.3.

This change fixes the jq logic to take the different diagnostics vesion into account. Note that other places in the pipeline already treat the diagnostics image special, this was just a missed case. The change also removes the `set` command from `release-tools.sh` to make local debugging eaiser, since that script is meant to be dot-sourced. Finally, I renamed the `release-core-images` to `refresh-core-images` so it isn't confused with the regular image release pipeline.

## Azure IoT Edge PR checklist: