Fix metrics collector release in main (#6834)
Azure Pipelines recently had a bug where they changed the default checkout behavior to fetch depth = 1, which is more performant but can cause problems with merges. They rolled back the change, but in the meantime, I added `fetchDepth: 0` in many places to mitigate. Adding it to the images-linux.yaml template causes a problem with the metrics collector release pipeline. I'm not sure if the problem is another bug in Azure Pipelines or not, but regardless they've rolled back their original change so I'm able to revert images-linux.yaml.

I ran the metrics collector release pipeline and the Build CI pipeline to confirm that the problem is fixed and doesn't appear to cause regressions in other pipelines.

## Azure IoT Edge PR checklist: