Fix manifest image build (#6847)
Our Docker image build template (image-linux.yaml) was recently modified to optionally build the manifest list too, as an optimization. But due to a copy/paste error, the template always builds the manifest list for Edge Agent instead of honoring the `manifestTemplate` param. This change fixes that problem. Note that currently, this feature is only being used for the auto-refresh builds.

I confirmed in a test environment that these changes have the desired effect.