Fix manifest image build, pt. II (#6849)
This is effectively a missed change from #6847. In the Docker image build template, there are two other arguments to `buildManifest.sh` that should be drawing from the parameters but are not. This change updates the template to use the `namespace` and `version` parameters. Without this fix, we fail to build the diagnostics module's manifest list when its version doesn't match the latest core runtime modules.

## Azure IoT Edge PR checklist: