Fix Metrics Collector release (#6920)
This change works around a problem that is specific to some of our release pipelines, which, prior to publishing the images, test arm32v7 Docker images on arm64v8 hosts to avoid using Raspberry Pis.

The problem, which cropped up recently when Docker changed its manifest format to support provenance attestation, is this: our arch-specific Docker images used to be manifests but now they're manifest lists with two entries, one pointing to the arch-specific manifest and the other pointing to provenance data. Because of this change, when the test agent's Docker daemon pulls the image, it sees the manifest list and attempts to select the appropriate manifest by matching a platform in the list to the host platform. Since we're testing arm32v7 images on arm64v8 hosts, 'arm/v7' != 'arm64/v8' and the host device can't pull the image.

The workaround is to extract the pointer to the arch-specific manifest from the list (sha256 hash) and refer to that instead.

To test, I made similar logic changes in the end-to-end test pipeline and confirmed that it worked as desired.

## Azure IoT Edge PR checklist: