Use newer tooling to copy image between registries (#6922)
Our script to copy an image from one registry to another uses the push/tag/pull pattern, which works fine in simple scenarios. But with recent changes to the structure of our Docker images (specifically, our arch-specific images are now technically multi-arch images with a single architecture in them, because of the way Docker now embeds provenance attestation information in images), it is better to use newer tooling that seamlessly handles all the variations.

This change replaces the docker pull, tag, and push calls in moveImage.sh with a single call to `docker buildx imagetools create`.

To test, I ran the script locally on my machine and confirmed that an image was copied between two of our internal registries.