Align to pre-1.4.9 Docker image structure (#6982)
Cherry-pick 0bae6b8a76d6756c33ed89c2361e34e276e41709. The only difference between this and the source commit in the main branch is that there's no Metrics Collector module (and associated release pipelines) in this branch.

To test, I ran my changes through the CI Build pipeline and confirmed that the resulting images passed the end-to-end tests.

I also set up the release pipelines in a test environment, and confirmed that the pipelines run and produce the expected images:
- API Proxy stage/publish
- Core images stage/publish
- Core images auto-refresh

## Azure IoT Edge PR checklist: