Fix test JSON and RHEL package build (#7020)
The JSON config we deploy in our ISA95 pipeline has an invalid entry--Edge agent's `https_proxy` environment variable should be embedded inside an `env` object, but it isn't. I recently upgraded the azure-iot extension to Azure CLI on one of the test agents, and newer versions of the extension validate the JSON config before sending it to IoT Hub, so the tests started failing. To test, I confirmed that the ISA95 smoke test pipeline passes with this change.

This also fixes a problem that cropped up yesterday in our RHEL 9 packages builds. We build inside an RHEL 9 container, and git received a major update in the container's backing image which caused a `git archive` call to fail because the container's user (uid 0) is different from the mounted source folder's owner (uid 1000). I added a git config update that skips the folder ownership check and allows the build to continue. To test, I confirmed that the CI Build pipeline now builds RHEL 9 packages.

## Azure IoT Edge PR checklist: