Support new image structure in auto-detect pipeline (#7023)
Cherry-pick a0f89a381d025cdaa38ff02ff495208d1ba50674.

To test, I ran the auto-detect pipeline and core images auto-refresh pipeline in a test environment and confirmed they behave as expected. I didn't go as far as forcing an image update; the auto-detect pipeline simply ran through the logic and determined there were no images to update. The core images auto-refesh pipeline was triggered as expected and quit when it determined there was no work to be done. This scenario required the least effort and tested most of the changes. Via inspection, I have reasonable confidence in the remaining changes. I also tested all of the jq statements in isolation, to ensure they parse and/or generate the expected JSON.

## Azure IoT Edge PR checklist: