Fix core refresh pipeline (#7032)
We currently have two auto-refresh pipelines
- Core images (edgeAgent, edgeHub, tempSensor, diagnostics) operates out of the release/1.4 branch
- Metrics collector operates out of main

Because of this separation, the two pipelines are able to operate independently, with one exception: they both need to update `product-versions.json` in the Azure/azure-iotedge repo.

Both pipelines were triggered yesterday at roughly the same time due to an update in the .NET Alpine base images. The metrics collector pipeline updated `product-versions.json` first, so by the time the core images pipeline tried to update it, it was no longer updating the head commit and the push failed to merge.

This change fixes the problem by syncing the latest changes before updating the JSON file. I suppose we could also add some retry logic in case a commit goes in during the brief period that the task is running, but that's very unlikely.

I also fixed a minor bug where the GitHub release pages should have "LTS" in the title but didn't.

To test, I ran the isolated changes in a local bash shell to ensure they work as expected. I didn't take the time to set up a full test environment. Worst case scenario is that the release pipeline will fail (like it already did last night), and the few remaining steps would need to be completed manually.

## Azure IoT Edge PR checklist: