Enable CodeQL in Build CI pipeline (#7059)
Per Microsoft policy, we need to run CodeQL static analysis regularly in our builds. It will run automatically in main if we simply set a variable on the job where the actual build occurs. This change does that. It also sets up a weekly scheduled run, in part to make sure that CodeQL runs on a regular basis. Note this PR sets up the schedule for main and release/1.4 branches, but it will need to be cherry-picked to the release/1.4 branch for the schedule to take effect there.

I also made an unrelated change: I removed the release/1.1 branch from the connectivity pipeline's weekly schedule, as that version of the product has been out of support for six months.

To test, I ran the Build CI pipeline against this PR and confirmed that it passes. CodeQL will run when it gets merged to main.