Enable 1ES-hosted Ubuntu 22.04 amd64 agent (#7064)
We tried to enable 1ES-hosted Ubuntu 22.04 agents (amd64 and arm64) in the end-to-end test pipeline a few months back, but we had to revert to custom agents because 1ES support for Ubuntu 22.04 wasn't quite there yet. Problems with amd64 have been resolved, so this change reintroduces the Ubuntu 22.04 amd64 1ES-hosted agent to the pipeline. This will allow us to decommision the corresponding custom agent and shed the associated overhead. Ubuntu 22.04 arm64 is still not ready, so we'll continue to use a custom agent there.

To test, I verified that the end-to-end test pipeline successfully uses a 1ES-hosted Ubuntu 22.04 amd64 agent.

## Azure IoT Edge PR checklist: