Remove Java header comment (#7069)
The CodeQL analayzer for our GitHub repo is detecting one of our C# files (edge-util/src/es6numberserializer/NumberFastDToABuilder.cs) as Java because a comment in the file's header overrides the heuristic. This change removes the comment so that the file is correctly identified as C#.

To test, I ran the [github-linguist](https://github.com/github-linguist/linguist) tool locally (this is the tool CodeQL uses) and confirmed that, with this change, no Java files are detected.

BEFORE:
```bash
azureuser@damonb-amd64-1-vm:~/iotedge$ github-linguist
74.23%  9531426    C#
21.50%  2761148    Rust
3.58%   459096     Shell
0.37%   47878      PowerShell
0.12%   15563      Dockerfile
0.11%   13615      Makefile
0.04%   5689       ANTLR
0.03%   4142       Java
0.02%   2536       Roff
```
AFTER:
```bash
~/iotedge$ github-linguist
74.26%  9535490    C#
21.50%  2761148    Rust
3.58%   459096     Shell
0.37%   47878      PowerShell
0.12%   15563      Dockerfile
0.11%   13615      Makefile
0.04%   5689       ANTLR
0.02%   2536       Roff
```

## Azure IoT Edge PR checklist: