Fix ISA95 smoke test pipeline (#7099)
In the setup scripts for the ISA95 smoke test pipeline, an Azure CLI command makes a bad assumption about the returned JSON, causing it to fail. The run does not abort at that point, but later when the corresponding cleanup scripts attempt to undo the earlier setup operation, that call fails and aborts the run with an error.

This change (1) forces the run to abort if the setup commands fail, and (2) fixes the setup command so that it can handle the more complex JSON result we're seeing. I also did some clean up and formatting of the command syntax for consistency.

To test, I ran the ISA95 Smoke Test pipeline, confirmed that it passed, and that the lock/unlock network scripts behaved as expected.

## Azure IoT Edge PR checklist: