Remove socket path if it is a directory instead of a file (#7088)
Updates edgelet's iot-identity-service dependency to the commit that added the fix for a bug causing socket file removal to fail when attempted on a directory instead of a file. See https://github.com/Azure/iot-identity-service/pull/544 for more details.

I tested the aziot-edge package built from this branch on an Ubuntu 20.04 VM to confirm the fix works.

## Azure IoT Edge PR checklist: