iotedge: default path to edged should be parsed from environment variable (#7154)
For non-standard installations of Azure IoT Edge (e.g., where the path to 'edged' is in a different location), this hardcoded path of the 'edged' binary may or may not be present. This change enables package builders for non-standard installations to wrap the binary with the given environment variable. One example would be the Linux distribution NixOS, where all binaries are stored under `/nix/store/...`. Since this change doesn't affect standard installations (e.g., when the environment variable is not set, it defaults), the benefits outweigh the change.