Only run CodeQL in one job of one pipeline (#7164)
In an earlier change (see #7059), we enabled CodeQL in one job of one pipeline, assuming it was off by default otherwise. It turns out that CodeQL is _on_ by default, so we need to disable it at the top of all pipelines. Then the one job of the one pipeline will enable it, and everything should work the way we expect.

Since CodeQL is automatically injected by policies within Microsoft, running the pipelines against this PR won't tell me anything. We may need to make further adjustments after this change is merged.

## Azure IoT Edge PR checklist: