Upgrade Functions sample to remove dependency on .NET Core 3.1 (#7167)
The Azure Functions module sample hasn't been updated in a while. It still depends on a .NET Core 3.1-based version of the Microsoft.NET.Sdk.Functions package. To remove our dependency on .NET Core 3.1 (which is out of support), this change upgrades the package to the latest version, which is based on .NET 6.0.

To test, I ran CI Build to get the bits, and then ran the end-to-end test pipeline. The test "TempFilterFunc" tests basic functionality of this sample.

## Azure IoT Edge PR checklist: