Don't override source images in Dockerfiles (#7169)
This change effectively reverts the work done in #6495. Since this is an OSS project, we have since learned that there are other ways to meet the relevant security requirements. The outcome is that the api-proxy module will be based on community Alpine (amd64) and nginx (arm32, arm64) images.

To test, I ran the CI build and confirmed that the API proxy image is built correctly. I also ran the nested end-to-end tests to confirm api-proxy module functionality.

## Azure IoT Edge PR checklist: