Don't `docker login` from IotEdgeCheck test (#7186)
The `iotedge check` end-to-end test was issuing a `docker login` command to ensure that the diagnostics image could be pulled even if it came from a different registry than Edge Agent and Edge Hub. That code has the potential to log docker credentials in the test output, so this change removes it. Also, while we may have needed to test with a diagnostics image from some other registry in the past, it's not necessary today. With these changes, the test simply assumes that the user running the tests has already given docker permissions to pull the diagnostics image before running the test.

The change also updates the test name to be consistent with the naming convention of other end-to-end tests. It also removes an unused variable discovered in the test infrastructure code.

To test, I ran the end-to-end tests pipeline and confirmed that the `iotedge check` test behaves as expected.

## Azure IoT Edge PR checklist: