Change API Proxy module build to use Debian 12 base image for ARM images. (#7209)
The base image of the module container uses `arm32v7/nginx` and `arm64v8/nginx` which are based on Debian 12, which uses OpenSSL v3. Before this change, we were building the binary itself in an Ubuntu 20.04 container so it would link to OpenSSL v1.1 instead and fail to run inside the Debian 12 containers.

This change fixes the build script to use a Debian 12 base image instead.