Add debug symbols to packages (#7211)
This change updates our build pipelines to build edgelet with full debug symbols, and then have the various packaging systems generate a separate symbols package for each distro. For the deb-based distros, the new package is something like `aziot-edge-dbgsym_{ver}_{arch}.deb` (`.ddeb` on Ubuntu). For the rpm-based distros, the new package is something like `aziot-edge-debuginfo-{ver}.{distro}.{arch}.rpm`.

This change also updates the test and release pipelines to handle the new packages. For the test pipelines, we don't want to copy/install the symbol packages. For the release pipeline, we _do_ want to sign and stage the symbol packages, but we only publish them to the GitHub release page for now, not to packages.microsoft.com.

To test, I ran the various pipelines to make sure the work as expected. I also manually installed IoT Edge, installed the symbol packages, and attached gdb to the running edgelet process to verify that it loads the stack frames.

## Azure IoT Edge PR checklist: