Update to the latest identity service commit ref (#7225)
We haven't updated iotedge@main's reference to iot-identity-service in a little while. Automation uses the reference to figure out which identity service package to download for test scenarios, but when the reference gets too old, GitHub Actions deletes the packages. The change updates the reference.

The checks for this PR will give the needed test coverage.

## Azure IoT Edge PR checklist: