Get service connection name from pipeline variable (#7241)
We changed a service connection in some of our pipelines and a hard-coded reference broke. This change updates the reference to a pipeline variable like all the other references in the repo.

Also, we have some special templates which are invoked in the end-to-end test pipeline from jobs that run on custom, stateful agents. The templates clean up build files and Docker images that may be hanging around from previous runs. Over time, we've switched several jobs to use stateless agents, but those jobs continue to invoke the templates. I removed references where they're no longer needed.

To test, I ran the end-to-end test pipeline and verified that it passed.

## Azure IoT Edge PR checklist: