aziot-edged: Only reprovision on Identity Service errors (#7246)
During startup, if aziot-edged encounters an error when obtaining the device identity, it will attempt to reprovision the device. Sometimes, if aziot-identityd has not fully started yet, aziot-edged will receive an OS "Connection refused" or similar error and attempt to reprovision.

Normally, the extra reprovision doesn't matter if the device is online. But in offline scenarios, attempting to reprovision will fail and put the device in a bad state.

This PR modifies the startup of aziot-edged so that only errors returned by Identity Service cause a reprovision. OS errors such as connection refused, permissions, etc. are ignored because they generally mean that Identity Service has not yet fully started.