Update changelog creation in auto-refresh image releases (#7248)
When the auto-refresh pipeline runs to update core module images (Edge Agent, Edge Hub, Simulated Temperature Sensor, and Diagnostics) following a base image update, we generate boilerplate text that goes into the [changelog](https://github.com/Azure/iotedge/blob/release/1.4/CHANGELOG.md) and the [GitHub release](https://github.com/Azure/azure-iotedge/releases). For no good reason, the message in the changelog is slightly different from the message in the release page--specifically, we prepend some extra information in the release page.

This change makes the two changelogs consistent. It will allow us to use a single pipeline for both manual and auto-refresh releases. I also updated the comment headers in a couple bash script functions to make it clear that they are usefull in both auto-refresh and manual release scenarios.

To test, I ran the release pipeline in a test environment and confirmed that the changelogs contain the expected information.

## Azure IoT Edge PR checklist: