Get identity service version from source instead of tag (#7280)
The Mariner arm64 build is currently failing. When it builds identity service bits, it tries to populate the version by getting the latest 1.5.* tag from the identity service repo. There are no 1.5.* tags yet, so it sends an empty version which causes `make` to fail.

The root cause is that we bumped our release version in code much sooner than we normally do (i.e., the period of time between bumping the version in code and actually publishing the bits is usually hours, but this time it's a couple weeks).

The fix is to extract the version from checked-in sources rather than git tag, since the tag isn't set until after we release.

To test, I ran the CI Build pipeline and confirmed that Mariner arm64 passes.

## Azure IoT Edge PR checklist: