Push release commit before pushing tag (#7290)
I noticed a problem today when .NET released new base images and our pipelines automatically triggered to update our core images. The pipeline created a release commit (locally, on the agent) with the updated changelog and versionInfo.json, but the commit failed to merge with the head of the branch, meaning we'll probably need to investigate the source of the merge problem, fix it manually, and re-run the pipeline--that's all fine. The problem is that the script _did_ manage to tag the release commit and push it upstream, so the pipeline re-run will fail to push the tag unless we manually clear it.

This change helps us prevent this problem in the future. It moves the tag push _after_ the merge commit push. I made the same change for both the core images release and the metrics collector image release.

## Azure IoT Edge PR checklist: