Remove temporary workaround in end-to-end test for iotedge check (#7292)
There was a period of a few weeks between when we updated the version in the code to 1.5.0 and when we actually shipped 1.5.0. During that window of time, one of the end-to-end tests would have failed so we added a workaround. Now that we've shipped 1.5.0, this change removes the workaround.

I ran the end-to-end tests and verified that the IotEdgeCheck test passes.

## Azure IoT Edge PR checklist: