Use SAS tokens to get an identity in E2E tests (#7309)
This PR introduces the use of short-lived SAS tokens to get an identity using the AzureCLI task. The use of shared keys to access the blob storage account is now discouraged, prompting this change.

**Testing notes:**

To test, I disabled shared key access on the storage account, then I ran the E2E check-in tests and confirmed that they're passing (and that the logs are uploaded).

![image](https://github.com/Azure/iotedge/assets/90283547/8dc0782e-2647-43db-8f07-0a39ab1d760a)


- [X] I have read the [contribution guidelines](https://github.com/azure/iotedge#contributing).