Upgrade alpine image for building rocksdb (#7325)
Docker has been throttling anonymous pull requests on their default docker.io registry for some time. We've avoided throttling by eliminated most of our references to base images on docker.io, but we still have one or two. Recently, one of the remaining references--the alpine image we use to build RocksDb in a container--has been getting throttled as well.

This change updates the Dockerfile we use to build RocksDb. It updates the base image to a mirrored version on mcr.microsoft.com.

To test, I ran the CI Build pipeline and confirmed that RocksDB is successfully built. I also ran the end-to-end test pipeline against the built binaries to confirm they behave as expected.

## Azure IoT Edge PR checklist: