Update cross builds to use default images (#7328)
With the introduction of throttling for anonymous pulls from docker.io, we've been systematically removing references to docker.io images. This change removes references to two custom images that we created years ago for cross-compiling edgelet for arm32v7 and arm64v8. By using the the `pre-build` key instead of `image`, we are able to invoke newer functionality in cross to create target environments just in time, based on the default images they maintain in the GitHub registry.

The "iotedged Checkin" check in this PR passes for Linux arm32v7 and arm64v8, confirming that the changes in this PR have the desired effect. This change was already made in the release/1.4 branch (see #7324).

Note I've also used this PR to delete a line from the edgelet dev guide that applies to CentOS 7, which is not supported in the main branch.

## Azure IoT Edge PR checklist: