Use SAS tokens to get an identity in Connectivity tests (#7313)
This PR introduces the use of short-lived SAS tokens to get an identity using the AzureCLI task. The use of shared keys to access the blob storage account is now discouraged, prompting this change.

We don't always upload logs to the storage account, only when we want to debug something. However, log upload is currently broken after the shared key on the storage account was disabled.

This is the manually triggered test run for this PR:

![image](https://github.com/Azure/iotedge/assets/90283547/a1d6b031-cdbb-4a45-b002-e581108bffa8)



- [X] I have read the [contribution guidelines](https://github.com/azure/iotedge#contributing).