Install snaps with devmode in end-to-end tests (#7343)
Following a recent fix in our snap packages, our end-to-end tests started to fail. The tests need to install the azure-iot-edge snap with --devmode, at least until we get approval to auto-connect the daemon-notify interface (in progress).

This change updates the install command in the end-to-end tests to prevent failures. I ran the end-to-end test pipeline and confirmed this fixes the failures.

## Azure IoT Edge PR checklist: