Fix end-to-end test failures on Raspberry Pi (#7349)
A recent change updated the some logic in the end-to-end tests to match the main branch, but an important detail was missed. The result is that end-to-end tests fail on Raspberry Pi in the release/1.4 branch because the tests don't recognize the device's operating system. This change corrects that problem.

I confirmed that the tests pass on Raspberry Pi with this change.

## Azure IoT Edge PR checklist: