Fix edgeAgent misreport Reported Properties (#7348)
TL;DR - When the cloud connection not working, the cached reported property should not be updated as it was not updated to the cloud.

Background: 
There is a bug when the edgeAgent starting and trying to download docker images for the other modules, it fails to download those module docker images (cloud connection failed), so the edgeAgent throw an exception (error).  Later, edgeAgent was able to download the module and continue with its operation which successfully apply the desired property to the device with version 42. Unfortunately, the interrupted edgeAgent misreported the version of the desired property as version 41 to IoTHub. 

## Azure IoT Edge PR checklist: